+++ CHANGES.md
@@ -1,8 +1,11 @@
-## HEAD
+## [v4.3.0]
+> Sep 30, 2016
 
 - Rename `onResponse` to `recv` and `enhanceFetch` to `enableRecv`. This makes the association more clear.
 - Deprecated redundant top-level `fetch` export. Use the global `fetch` function directly instead.
 - Added `debug` middleware, deprecated `requestInfo`.
+
+[v4.3.0]: https://github.com/mjackson/http-client/compare/v4.2.0...v4.3.0
 
 ## [v4.2.0]
 > Sep 29, 2016